+++ packages/bodiless-filtering/src/FilterByGroup/FilterByGroupStore.ts
@@ -12,25 +12,32 @@
  * limitations under the License.
  */
 
-import { useState, useEffect } from 'react';
+import {
+  useState, useEffect, useRef, useCallback,
+} from 'react';
+import { TagType } from '@bodiless/core';
+import { FilterTagType } from './types';
 
 const TAG_ANY_KEY = 'any';
 
-class Tag {
+class Tag implements FilterTagType {
   id: string;
 
   categoryId: string;
 
   name: string;
 
-  constructor(id: string, name: string, categoryId?: string) {
+  categoryName: string;
+
+  constructor(id: string, name: string, categoryId?: string, categoryName?: string) {
     this.id = id;
-    this.categoryId = categoryId || '';
     this.name = name;
+    this.categoryId = categoryId || '';
+    this.categoryName = categoryName || '';
   }
 
-  isEqual(tag: Tag) {
-    return tag.id === this.id && tag.categoryId === this.categoryId;
+  isEqual(tag: TagType) {
+    return tag.id === this.id && (tag as FilterTagType).categoryName === this.categoryName;
   }
 }
 
@@ -42,13 +49,14 @@ const readTagsFromQueryParams = () => {
   if (typeof window === 'undefined') return [];
   const tags: Tag[] = [];
   const params = new URLSearchParams(window.location.search);
-  params.forEach((tagId, categoryId) => {
-    tags.push(new Tag(tagId, '', categoryId));
+  params.forEach((tag, categoryId) => {
+    const [tagId, tagName = 'N/A', categoryName = 'N/A'] = tag.split('~');
+    tags.push(new Tag(tagId, tagName, categoryId, categoryName));
   });
   return tags;
 };
 
-const updateUrlQueryParams = (tags: Tag[]) => {
+const updateUrlQueryParams = (tags: FilterTagType[]) => {
   if (typeof window === 'undefined') return;
   const {
     protocol,
@@ -57,52 +65,78 @@ const updateUrlQueryParams = (tags: Tag[]) => {
   } = window.location;
   const queryParams = new URLSearchParams();
   tags.forEach(tag => {
-    queryParams.append(tag.categoryId || '', tag.id);
+    const {
+      categoryId, id, name, categoryName,
+    } = tag;
+    queryParams.append(categoryId || '', `${id}~${name}~${categoryName}`);
   });
   const query = tags.length > 0 ? `?${queryParams}` : '';
   const newurl = `${protocol}//${host}${pathname}${query}`;
   window.history.pushState({ path: newurl }, '', newurl);
 };
 
+const useStateCallback = (initialState: any) => {
+  const [state, setState] = useState(initialState);
+  const cbRef = useRef<Function | null>(null); // init mutable ref container for callbacks
+
+  const setStateCallback = useCallback((newState, cb) => {
+    cbRef.current = cb; // store current, passed callback in ref
+    setState(newState);
+  }, []); // keep object reference stable, exactly like `useState`
+
+  useEffect(() => {
+    // cb.current is `null` on initial render,
+    // so we only invoke callback on state *updates*
+    if (cbRef.current) {
+      cbRef.current(state);
+      cbRef.current = null; // reset callback after execution
+    }
+  }, [state]);
+
+  return [state, setStateCallback];
+};
+
 const useFilterByGroupStore = (settings: FilterByGroupStoreSettings) => {
-  const [selectedTags, setSelectedTags] = useState<Tag[]>([]);
+  const [selectedTags, setSelectedTags] = useStateCallback([]);
+  const [filtersInitialized, setFiltersInitialized] = useState<boolean>(false);
 
   useEffect(() => {
     const tags = readTagsFromQueryParams();
     setSelectedTags(tags);
+    setFiltersInitialized(true);
   }, []);
 
   const { multipleAllowedTags = false } = settings;
 
-  const updateSelectedTags = (tags: Tag[]) => {
+  const updateSelectedTags = (tags: Tag[], callback?: Function) => {
     updateUrlQueryParams(tags);
-    setSelectedTags(tags);
+    setSelectedTags(tags, callback);
   };
 
-  const selectTag = (tag: Tag) => {
+  const selectTag = (tag: FilterTagType, callback?: Function) => {
     updateSelectedTags([
       ...(
         multipleAllowedTags
           ? selectedTags
-          : selectedTags.filter(tag$ => tag.categoryId !== tag$.categoryId)
+          : selectedTags.filter((tag$: FilterTagType) => tag.categoryId !== tag$.categoryId)
       ),
       tag,
-    ]);
+    ], callback);
   };
 
-  const unSelectTag = (tag: Tag) => {
+  const unSelectTag = (tag: Tag, callback?: Function) => {
     updateSelectedTags([
-      ...selectedTags.filter(tag$ => !tag.isEqual(tag$)),
-    ]);
+      ...selectedTags.filter((tag$: Tag) => !tag.isEqual(tag$)),
+    ], callback);
   };
 
   const isTagSelected = (tag: Tag) => {
     if (!multipleAllowedTags && tag.id === TAG_ANY_KEY) {
       // For radios, return true for ANY tag if no other tags are selected.
-      const tagsInCategory = selectedTags.filter(t => t.categoryId === tag.categoryId);
+      const tagsInCategory = selectedTags.filter((t: Tag) => t.categoryId === tag.categoryId);
       if (tagsInCategory.length === 0) return true;
     }
-    return selectedTags.find(tag$ => tag.isEqual(tag$)) !== undefined;
+    return selectedTags.find((tag$: Tag) => tag.isEqual(tag$)) !== undefined;
   };
 
   const getSelectedTags = () => selectedTags;
@@ -115,6 +149,7 @@ const useFilterByGroupStore = (settings: FilterByGroupStoreSettings) => {
     getSelectedTags,
     isTagSelected,
     clearSelectedTags,
+    filtersInitialized,
   };
 };