+++ packages/bodiless-filtering/src/index.ts
@@ -1,1 +1,3 @@
 export * from './FilterByGroup';
+export * from './TagButton';
+export * from './Taggable/asTaggableItem';